Prevent leaking of internal information to unprivileged users
Currently, if `SAVEQUERIES` is defined and true (e.g. you installed + activated http://wordpress.org/plugins/query-monitor/), then Sermon Browser will cause the output of information about all database queries (and file paths) in the HTML source, to all users. (I discovered this by examining my page cache files). As well as the information leakage, this was adding 680KB of size to the HTML source for me.

The mere presence of that constant is not sufficient to indicate that the site owner wishes debugging information to be made available to all visitors. The change in this branch restricts its display to logged-in + privileged users who have turned on debugging.